+++ crates/db-storage/tests/get_events.rs
@@ -17,7 +17,7 @@ fn make_event(
     conn: &mut DbConnection,
     user_id: UserId,
     room_id: RoomId,
-    hour: u32,
+    hour: Option<u32>,
     is_adhoc: bool,
 ) -> Event {
     NewEvent {
@@ -26,13 +26,13 @@ fn make_event(
         room: room_id,
         created_by: user_id,
         updated_by: user_id,
-        is_time_independent: false,
+        is_time_independent: hour.is_none(),
         is_all_day: Some(false),
-        starts_at: Some(Tz::UTC.ymd(2020, 1, 1).and_hms(hour, 0, 0)),
-        starts_at_tz: Some(TimeZone(Tz::UTC)),
-        ends_at: Some(Tz::UTC.ymd(2020, 1, 1).and_hms(hour, 0, 0)),
-        ends_at_tz: Some(TimeZone(Tz::UTC)),
-        duration_secs: Some(0),
+        starts_at: hour.map(|h| Tz::UTC.ymd(2020, 1, 1).and_hms(h, 0, 0)),
+        starts_at_tz: hour.map(|_| TimeZone(Tz::UTC)),
+        ends_at: hour.map(|h| Tz::UTC.ymd(2020, 1, 1).and_hms(h, 0, 0)),
+        ends_at_tz: hour.map(|_| TimeZone(Tz::UTC)),
+        duration_secs: hour.map(|_| 0),
         is_recurring: Some(false),
         recurrence_pattern: None,
         is_adhoc,
@@ -89,18 +89,18 @@ async fn test() {
 
     // first two events, first on on hour 2 then 1.
     // This tests the ordering of comparison of times (e.g. starts_at, then created_at)
-    let event2 = make_event(&mut conn, user.id, room.id, 2, true);
-    let event1 = make_event(&mut conn, user.id, room.id, 1, true);
+    let event2 = make_event(&mut conn, user.id, room.id, Some(2), true);
+    let event1 = make_event(&mut conn, user.id, room.id, Some(1), true);
 
     // this event should come last because starts_at is largest
-    let event8 = make_event(&mut conn, user.id, room.id, 10, true);
+    let event8 = make_event(&mut conn, user.id, room.id, Some(10), true);
 
     // Test that created_at is being honored if starts_at is equal
-    let event3 = make_event(&mut conn, user.id, room.id, 3, true);
-    let event4 = make_event(&mut conn, user.id, room.id, 3, true);
-    let event5 = make_event(&mut conn, user.id, room.id, 3, true);
-    let event6 = make_event(&mut conn, user.id, room.id, 3, true);
-    let event7 = make_event(&mut conn, user.id, room.id, 3, true);
+    let event3 = make_event(&mut conn, user.id, room.id, Some(3), true);
+    let event4 = make_event(&mut conn, user.id, room.id, Some(3), true);
+    let event5 = make_event(&mut conn, user.id, room.id, Some(3), true);
+    let event6 = make_event(&mut conn, user.id, room.id, Some(3), true);
+    let event7 = make_event(&mut conn, user.id, room.id, Some(3), true);
 
     {
         // Test cursor
@@ -111,6 +111,7 @@ async fn test() {
             user.id,
             false,
             vec![],
+            None,
             None,
             None,
             None,
@@ -136,6 +137,7 @@ async fn test() {
             None,
             None,
             None,
+            None,
             Some(cursor),
             2,
         )
@@ -157,6 +159,7 @@ async fn test() {
             None,
             None,
             None,
+            None,
             Some(cursor),
             2,
         )
@@ -175,6 +178,7 @@ async fn test() {
             user.id,
             false,
             vec![],
+            None,
             None,
             None,
             None,
@@ -200,6 +204,7 @@ async fn test() {
             None,
             None,
             None,
+            None,
             100,
         )
         .unwrap();
@@ -216,6 +221,7 @@ async fn test() {
             vec![],
             None,
             Some(Utc.ymd(2020, 1, 1).and_hms(5, 0, 0)),
+            None,
             None,
             None,
             100,
@@ -240,6 +246,7 @@ async fn test() {
             vec![],
             Some(Utc.ymd(2020, 1, 1).and_hms(3, 0, 0)),
             Some(Utc.ymd(2020, 1, 1).and_hms(3, 0, 0)),
+            None,
             None,
             None,
             100,
@@ -272,10 +279,10 @@ async fn get_events_invite_filter() {
     .insert(&mut conn)
     .unwrap();
 
-    let accept_event = make_event(&mut conn, inviter.id, room.id, 1, true);
-    let decline_event = make_event(&mut conn, inviter.id, room.id, 1, true);
-    let tentative_event = make_event(&mut conn, inviter.id, room.id, 1, true);
-    let pending_event = make_event(&mut conn, inviter.id, room.id, 1, true);
+    let accept_event = make_event(&mut conn, inviter.id, room.id, Some(1), true);
+    let decline_event = make_event(&mut conn, inviter.id, room.id, Some(1), true);
+    let tentative_event = make_event(&mut conn, inviter.id, room.id, Some(1), true);
+    let pending_event = make_event(&mut conn, inviter.id, room.id, Some(1), true);
 
     // Check that the creator of the events gets created events when filtering by `Accepted` invite status
     let all_events = Event::get_all_for_user_paginated(
@@ -283,6 +290,7 @@ async fn get_events_invite_filter() {
         inviter.id,
         false,
         vec![EventInviteStatus::Accepted],
+        None,
         None,
         None,
         None,
@@ -311,6 +319,7 @@ async fn get_events_invite_filter() {
         inviter.id,
         false,
         vec![EventInviteStatus::Declined],
+        None,
         None,
         None,
         None,
@@ -371,6 +380,7 @@ async fn get_events_invite_filter() {
         None,
         None,
         None,
+        None,
         100,
     )
     .unwrap();
@@ -386,6 +396,7 @@ async fn get_events_invite_filter() {
         invitee.id,
         false,
         vec![EventInviteStatus::Declined],
+        None,
         None,
         None,
         None,
@@ -409,6 +420,7 @@ async fn get_events_invite_filter() {
         None,
         None,
         None,
+        None,
         100,
     )
     .unwrap();
@@ -428,6 +440,7 @@ async fn get_events_invite_filter() {
         None,
         None,
         None,
+        None,
         100,
     )
     .unwrap();
@@ -443,6 +456,7 @@ async fn get_events_invite_filter() {
         invitee.id,
         false,
         vec![],
+        None,
         None,
         None,
         None,
@@ -486,7 +500,7 @@ async fn get_event_invites() {
     .unwrap();
 
     // EVENT 1 MIT JEEZ LOUISE AND GERHARD
-    let event1 = make_event(&mut conn, ferdinand.id, room.id, 1, true);
+    let event1 = make_event(&mut conn, ferdinand.id, room.id, Some(1), true);
 
     NewEventInvite {
         event_id: event1.id,
@@ -507,7 +521,7 @@ async fn get_event_invites() {
     .unwrap();
 
     // EVENT 2 MIT JEEZ LOUSE UND FERDINAND
-    let event2 = make_event(&mut conn, gerhard.id, room.id, 1, true);
+    let event2 = make_event(&mut conn, gerhard.id, room.id, Some(1), true);
 
     NewEventInvite {
         event_id: event2.id,
@@ -576,17 +590,18 @@ async fn get_event_adhoc() {
     .insert(&mut conn)
     .unwrap();
 
-    let event1 = make_event(&mut conn, user.id, room.id, 1, true);
-    let event2 = make_event(&mut conn, user.id, room.id, 1, false);
-    let event3 = make_event(&mut conn, user.id, room.id, 1, false);
-    let event4 = make_event(&mut conn, user.id, room.id, 1, true);
-    let event5 = make_event(&mut conn, user.id, room.id, 1, true);
+    let event1 = make_event(&mut conn, user.id, room.id, Some(1), true);
+    let event2 = make_event(&mut conn, user.id, room.id, Some(1), false);
+    let event3 = make_event(&mut conn, user.id, room.id, Some(1), false);
+    let event4 = make_event(&mut conn, user.id, room.id, Some(1), true);
+    let event5 = make_event(&mut conn, user.id, room.id, Some(1), true);
 
     let all = Event::get_all_for_user_paginated(
         &mut conn,
         user.id,
         false,
         vec![],
+        None,
         None,
         None,
         None,
@@ -611,6 +626,7 @@ async fn get_event_adhoc() {
         None,
         Some(true),
         None,
+        None,
         10,
     )
     .unwrap();
@@ -628,10 +644,109 @@ async fn get_event_adhoc() {
         None,
         Some(false),
         None,
+        None,
         10,
     )
     .unwrap();
     assert_eq!(non_adhoc.len(), 2);
     assert_eq!(non_adhoc[0].0, event2);
     assert_eq!(non_adhoc[1].0, event3);
+}
+
+#[tokio::test]
+#[serial]
+async fn get_event_time_independent() {
+    let db_ctx = test_util::database::DatabaseContext::new(true).await;
+
+    let mut conn = db_ctx.db.get_conn().unwrap();
+
+    let (user, _) = NewUserWithGroups {
+        new_user: NewUser {
+            email: "test@example.org".into(),
+            title: "".into(),
+            firstname: "Test".into(),
+            lastname: "Tester".into(),
+            id_token_exp: 0,
+            language: "".into(),
+            display_name: "Test Tester".into(),
+            oidc_sub: "testtestersoidcsub".into(),
+            oidc_issuer: "".into(),
+            phone: None,
+        },
+        groups: vec![],
+    }
+    .insert(&mut conn)
+    .unwrap();
+
+    let room = NewRoom {
+        created_by: user.id,
+        password: None,
+        waiting_room: false,
+    }
+    .insert(&mut conn)
+    .unwrap();
+
+    let event1 = make_event(&mut conn, user.id, room.id, None, false);
+    let event2 = make_event(&mut conn, user.id, room.id, None, false);
+    let event3 = make_event(&mut conn, user.id, room.id, Some(1), false);
+    let event4 = make_event(&mut conn, user.id, room.id, None, false);
+    let event5 = make_event(&mut conn, user.id, room.id, Some(2), false);
+
+    let all = Event::get_all_for_user_paginated(
+        &mut conn,
+        user.id,
+        false,
+        vec![],
+        None,
+        None,
+        None,
+        None,
+        None,
+        10,
+    )
+    .unwrap();
+
+    // different order than creation order, because events without
+    // starts_at field are sorted first
+    assert_eq!(all.len(), 5);
+    assert_eq!(all[0].0, event1);
+    assert_eq!(all[1].0, event2);
+    assert_eq!(all[2].0, event4);
+    assert_eq!(all[3].0, event3);
+    assert_eq!(all[4].0, event5);
+
+    let time_independent = Event::get_all_for_user_paginated(
+        &mut conn,
+        user.id,
+        false,
+        vec![],
+        None,
+        None,
+        None,
+        Some(true),
+        None,
+        10,
+    )
+    .unwrap();
+    assert_eq!(time_independent.len(), 3);
+    assert_eq!(time_independent[0].0, event1);
+    assert_eq!(time_independent[1].0, event2);
+    assert_eq!(time_independent[2].0, event4);
+
+    let time_dependent = Event::get_all_for_user_paginated(
+        &mut conn,
+        user.id,
+        false,
+        vec![],
+        None,
+        None,
+        None,
+        Some(false),
+        None,
+        10,
+    )
+    .unwrap();
+    assert_eq!(time_dependent.len(), 2);
+    assert_eq!(time_dependent[0].0, event3);
+    assert_eq!(time_dependent[1].0, event5);
 }